+++ src/components/FinesseAbilitiesSection.tsx
@@ -24,13 +24,25 @@ export function FinesseAbilitiesSection({ dex, finesse_points, max_finesse_point
           <Paper p="xs" withBorder style={STYLES.CARD_BACKGROUND}>
             <Text size="sm" fw={500} style={{ color: COLORS.FINESSE_ABILITY }}>Sneak Attack</Text>
             <Text size="xs" style={STYLES.DESCRIPTION_TEXT}>
-              Roll {max_finesse_points}d8 extra damage when attacking with advantage or when target is engaged with an ally. Costs 1 finesse point per use.
+              Roll {finesse_points}d8 extra damage when attacking with advantage or when target is engaged with an ally. Costs 1 finesse point per use.
+            </Text>
+          </Paper>
+          <Paper p="xs" withBorder style={STYLES.CARD_BACKGROUND}>
+            <Text size="sm" fw={500} style={{ color: COLORS.FINESSE_ABILITY }}>Assassination</Text>
+            <Text size="xs" style={STYLES.DESCRIPTION_TEXT}>
+              Roll {finesse_points}d8 extra damage when attacking an unaware enemy who has taken no turns in combat. Automatically hits critically if the attack roll succeeds. Costs no finesse points.
             </Text>
           </Paper>
           <Paper p="xs" withBorder style={STYLES.CARD_BACKGROUND}>
             <Text size="sm" fw={500} style={{ color: COLORS.FINESSE_ABILITY }}>Enhanced Hide</Text>
             <Text size="xs" style={STYLES.DESCRIPTION_TEXT}>
-              When hiding in light armor or no armor, double your level bonus to stealth rolls
+              When hiding, double your level bonus to stealth rolls.
+            </Text>
+          </Paper>
+          <Paper p="xs" withBorder style={STYLES.CARD_BACKGROUND}>
+            <Text size="sm" fw={500} style={{ color: COLORS.FINESSE_ABILITY }}>Sleight of Hand</Text>
+            <Text size="xs" style={STYLES.DESCRIPTION_TEXT}>
+			  Double the your level bonus when performing a sleight-of-hand check. 
             </Text>
           </Paper>
         </Stack>